Refactor poker analyzer screen into modular widgets
