+++ lib/screens/poker_analyzer/playback_controls_widget.dart
@@ -1,4 +1,4 @@
-part of '../poker_analyzer_screen.dart';
+import 'package:flutter/material.dart';
 
 class _SaveLoadControlsSection extends StatelessWidget {
   final VoidCallback onSave;
